feat(pkg/crypto): move `crypto` package to `pkg`
