--- crypto/secp256k1.go
@@ -1,163 +1,0 @@
-package crypto
-
-import (
-	"crypto/sha512"
-	"encoding/hex"
-	"errors"
-	"math/big"
-	"strings"
-
-	"github.com/btcsuite/btcd/btcec/v2"
-	"github.com/decred/dcrd/dcrec/secp256k1/v4"
-	ecdsa "github.com/decred/dcrd/dcrec/secp256k1/v4/ecdsa"
-)
-
-const (
-	// SECP256K1 prefix - value is 0
-	secp256K1Prefix = 0x00
-	// SECP256K1 family seed prefix - value is 33
-	secp256K1FamilySeedPrefix = 0x21
-)
-
-type secp256K1CryptoAlgorithm CryptoAlgorithm
-
-func SECP256K1() secp256K1CryptoAlgorithm {
-	return secp256K1CryptoAlgorithm{
-		prefix: secp256K1Prefix,
-		familySeedPrefix: secp256K1FamilySeedPrefix,
-	}
-}
-
-func (c secp256K1CryptoAlgorithm) Prefix() byte {
-	return c.prefix
-}
-
-func (c secp256K1CryptoAlgorithm) FamilySeedPrefix() byte {
-	return c.familySeedPrefix
-}
-
-func (c secp256K1CryptoAlgorithm) deriveScalar(bytes []byte, discrim *big.Int) *big.Int {
-
-	order := btcec.S256().N
-	for i := 0; i <= 0xffffffff; i++ {
-		hash := sha512.New()
-
-		hash.Write(bytes)
-
-		if discrim != nil {
-			discrimBytes := make([]byte, 4)
-			bytes[0] = byte(uint32(discrim.Uint64()))
-			bytes[1] = byte(uint32(discrim.Uint64()) >> 8)
-			bytes[2] = byte(uint32(discrim.Uint64()) >> 16)
-			bytes[3] = byte(uint32(discrim.Uint64()) >> 24)
-
-			hash.Write(discrimBytes)
-		}
-
-		shiftBytes := make([]byte, 4)
-		bytes[0] = byte(uint32(i))
-		bytes[1] = byte(uint32(i) >> 8)
-		bytes[2] = byte(uint32(i) >> 16)
-		bytes[3] = byte(uint32(i) >> 24)
-
-		hash.Write(shiftBytes)
-
-		key := new(big.Int).SetBytes(hash.Sum(nil)[:32])
-
-		if key.Cmp(big.NewInt(0)) > 0 && key.Cmp(order) < 0 {
-			return key
-		}
-	}
-	// This error is practically impossible to reach.
-	// The order of the curve describes the (finite) amount of points on the curve.
-	panic("impossible unicorn ;)")
-}
-
-func (c secp256K1CryptoAlgorithm) DeriveKeypair(seed []byte, validator bool) (string, string, error) {
-	curve := btcec.S256()
-	order := curve.N
-
-	privateGen := c.deriveScalar(seed, nil)
-
-	if validator {
-		return "", "", errors.New("validator keypair derivation not supported")
-	}
-
-	rootPrivateKey, _ := btcec.PrivKeyFromBytes(privateGen.Bytes())
-
-	derivatedScalar := c.deriveScalar(rootPrivateKey.PubKey().SerializeCompressed(), big.NewInt(0))
-	scalarWithPrivateGen := derivatedScalar.Add(derivatedScalar, privateGen)
-	privateKey := scalarWithPrivateGen.Mod(scalarWithPrivateGen, order)
-
-	privKeyBytes := privateKey.Bytes()
-	private := strings.ToUpper(hex.EncodeToString(privKeyBytes))
-
-	_, pubKey := btcec.PrivKeyFromBytes(privKeyBytes)
-
-	pubKeyBytes := pubKey.SerializeCompressed()
-
-	return "00" + private, strings.ToUpper(hex.EncodeToString(pubKeyBytes)), nil
-}
-
-func (c secp256K1CryptoAlgorithm) Sign(msg, privKey string) (string, error) {
-	if len(privKey) != 64 && len(privKey) != 66 {
-		return "", errors.New("invalid private key")
-	}
-	if len(msg) == 0 {
-		return "", errors.New("message is required")
-	}
-
-	if len(privKey) == 66 {
-		privKey = privKey[2:]
-	}
-	key, err := hex.DecodeString(privKey)
-	if err != nil {
-		return "", err
-	}
-
-	secpPrivKey := secp256k1.PrivKeyFromBytes(key)
-	sig := ecdsa.Sign(secpPrivKey, Sha512Half([]byte(msg)))
-
-	parsedSig, err := DERHexFromSig(sig.R().String(), sig.S().String())
-	if err != nil {
-		return "", err
-	}
-	return strings.ToUpper(parsedSig), nil
-}
-
-func (c secp256K1CryptoAlgorithm) Validate(msg, pubkey, sig string) bool {
-	// Decode the signature from DERHex to a hex string
-	r, s, err := DERHexToSig(sig)
-	if err != nil {
-		return false
-	}
-
-	// Convert r and s slices to [32]byte arrays
-	var rBytes, sBytes [32]byte
-
-	copy(rBytes[32-len(r):], r)
-	copy(sBytes[32-len(s):], s)
-
-	ecdsaR := &secp256k1.ModNScalar{}
-	ecdsaS := &secp256k1.ModNScalar{}
-
-	ecdsaR.SetBytes(&rBytes)
-	ecdsaS.SetBytes(&sBytes)
-
-	parsedSig := ecdsa.NewSignature(ecdsaR, ecdsaS)
-	// Hash the message
-	hash := Sha512Half([]byte(msg))
-
-	// Decode the pubkey from hex to a byte slice
-	pubkeyBytes, err := hex.DecodeString(pubkey)
-	if err != nil {
-		return false
-	}
-
-	// Verify the signature
-	pubKey, err := secp256k1.ParsePubKey(pubkeyBytes)
-	if err != nil {
-		return false
-	}
-	return parsedSig.Verify(hash, pubKey)
-}
