--- crypto/types.go
@@ -1,8 +1,0 @@
-package crypto
-
-type CryptoAlgorithm struct {
-	prefix                   uint8
-	familySeedPrefix         uint8
-}
-
-
