+++ pkg/crypto/secp256k1_test.go
@@ -1,16 +1,14 @@
-package keypairs
+package crypto
 
 import (
 	"errors"
 	"testing"
-
-	addresscodec "github.com/Peersyst/xrpl-go/address-codec"
 )
 
 func TestSecp256k1_deriveKeypair(t *testing.T) {
 	testCases := []struct {
 		name            string
-		seed            string
+		seedBytes       []byte
 		validator       bool
 		expectedPrivKey string
 		expectedPubKey  string
@@ -18,7 +16,7 @@ func TestSecp256k1_deriveKeypair(t *testing.T) {
 	}{
 		{
 			name:            "valid seed (1)",
-			seed:            "sntbkd2DsouBx8BAdJdi35p1HRw6h",
+			seedBytes:       []byte{229, 81, 182, 134, 131, 220, 192, 126, 133, 114, 150, 132, 140, 237, 222, 196},
 			validator:       false,
 			expectedPubKey:  "02950F4710101A25073BF37086D73FBBD00C7A6B0F91097D8F0BC6D268C400D56E",
 			expectedPrivKey: "00B167A9F3B9E60A4F93695713682C102438620AA1785C3AE635F53E5B6261071A",
@@ -26,7 +24,7 @@ func TestSecp256k1_deriveKeypair(t *testing.T) {
 		},
 		{
 			name:            "valid seed (2)",
-			seed:            "shSDdnXqsS7zAjbdWX86fT6H5oCxK",
+			seedBytes:       []byte{124, 228, 51, 247, 54, 54, 81, 51, 239, 86, 226, 187, 232, 20, 111, 163},
 			validator:       false,
 			expectedPubKey:  "031FBCFDD2EC6C2EDFBBA3866BDBAC28E5253C6A01FE9EFF8CAAE01871F009E837",
 			expectedPrivKey: "00A3D1513DBE784107428B363A1F8EAF1377AB63D4D137AB9E28E0BC614C71D8C0",
@@ -34,7 +32,7 @@ func TestSecp256k1_deriveKeypair(t *testing.T) {
 		},
 		{
 			name:            "validator set to true",
-			seed:            "shSDdnXqsS7zAjbdWX86fT6H5oCxK",
+			seedBytes:       []byte{124, 228, 51, 247, 54, 54, 81, 51, 239, 86, 226, 187, 232, 20, 111, 163},
 			validator:       true,
 			expectedPubKey:  "",
 			expectedPrivKey: "",
@@ -44,12 +42,7 @@ func TestSecp256k1_deriveKeypair(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			secp256k1 := secp256k1Alg{}
-			seedBytes, _, err := addresscodec.DecodeSeed(tc.seed)
-			if err != nil {
-				t.Fatal(err)
-			}
-			privKey, pubKey, err := secp256k1.deriveKeypair(seedBytes, tc.validator)
+			privKey, pubKey, err := SECP256K1().DeriveKeypair(tc.seedBytes, tc.validator)
 			if tc.expectedErr != nil {
 				if err == nil || err.Error() != tc.expectedErr.Error() {
 					t.Fatalf("expected error %v, got %v", tc.expectedErr, err)
@@ -144,8 +137,7 @@ func TestSecp256k1_sign(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			secp256k1 := secp256k1Alg{}
-			signature, err := secp256k1.sign(tc.message, tc.privKey)
+			signature, err := SECP256K1().Sign(tc.message, tc.privKey)
 
 			if signature != tc.expectedSignature {
 				t.Errorf("sign() returned %v, want %v", signature, tc.expectedSignature)
@@ -210,8 +202,7 @@ func TestSecp256k1_validate(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			secp256k1 := secp256k1Alg{}
-			isValid := secp256k1.validate(tc.message, tc.pubKey, tc.signature)
+			isValid := SECP256K1().Validate(tc.message, tc.pubKey, tc.signature)
 
 			if isValid != tc.wantValid {
 				t.Errorf("validate() = %v, want %v", isValid, tc.wantValid)
